Extract out some logic to runner package
